refactor(handler): merge response.go into responsestep.go
response.go had a single consumer (handler package) and its content — ACK/NACK body
construction and wire-format helpers — is logically part of the response-step layer.
Merge preV2Ack/Message/Response types, sendAck/sendNack/nack helpers, and their
tests into responsestep.go / responsestep_test.go; delete the now-empty response.go
and response_test.go files. No behaviour change.
