+++ core/module/handler/responsestep.go
@@ -2,7 +2,9 @@ package handler
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"net/http"
@@ -14,7 +16,161 @@ import (
 )
 
 // ---------------------------------------------------------------------------
-// ackSignerStep
+// Response helpers — ACK / NACK body construction and writing
+// ---------------------------------------------------------------------------
+
+// preV2Ack is the pre-v2 acknowledgment wrapper used for context.version < "2.0.0".
+// Wire format: {"ack":{"status":"ACK"}}
+type preV2Ack struct {
+	Status model.Status `json:"status"`
+}
+
+// preV2Message is the pre-v2 message envelope.
+// Wire format: {"message":{"ack":{"status":"ACK"},"error":{...}}}
+type preV2Message struct {
+	Ack   preV2Ack     `json:"ack"`
+	Error *model.Error `json:"error,omitempty"`
+}
+
+// preV2Response is the pre-v2 top-level response.
+type preV2Response struct {
+	Message preV2Message `json:"message"`
+}
+
+// sendAck sends a synchronous ACK response to the client.
+// For context.version "2.0.0" and later the response uses the v2 envelope:
+//
+//	{"message":{"status":"ACK","messageId":"<uuid>"}}
+//
+// All other versions use the pre-v2 envelope:
+//
+//	{"message":{"ack":{"status":"ACK"}}}
+func sendAck(ctx context.Context, w http.ResponseWriter) {
+	var data []byte
+	if isAtLeastV2(ctx) {
+		resp := &model.Response{
+			Message: model.Message{
+				Status:    model.StatusACK,
+				MessageID: msgID(ctx),
+			},
+		}
+		data, _ = json.Marshal(resp)
+	} else {
+		resp := &preV2Response{
+			Message: preV2Message{
+				Ack: preV2Ack{Status: model.StatusACK},
+			},
+		}
+		data, _ = json.Marshal(resp)
+	}
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	if _, err := w.Write(data); err != nil {
+		http.Error(w, "failed to write response", http.StatusInternalServerError)
+	}
+}
+
+// nackBecknError maps an error to its Beckn model.Error representation and HTTP
+// status code.
+func nackBecknError(ctx context.Context, err error) (*model.Error, int) {
+	var schemaErr *model.SchemaValidationErr
+	var signErr *model.SignValidationErr
+	var badReqErr *model.BadReqErr
+	var notFoundErr *model.NotFoundErr
+
+	switch {
+	case errors.As(err, &schemaErr):
+		return schemaErr.BecknError(), http.StatusBadRequest
+	case errors.As(err, &signErr):
+		return signErr.BecknError(), http.StatusUnauthorized
+	case errors.As(err, &badReqErr):
+		return badReqErr.BecknError(), http.StatusBadRequest
+	case errors.As(err, &notFoundErr):
+		return notFoundErr.BecknError(), http.StatusNotFound
+	default:
+		return internalServerError(ctx), http.StatusInternalServerError
+	}
+}
+
+// nackBodyBytes returns the serialised NACK response body for the given Beckn
+// error. The output is identical to the bytes that nack() writes to the wire.
+func nackBodyBytes(ctx context.Context, becknErr *model.Error) []byte {
+	var data []byte
+	if isAtLeastV2(ctx) {
+		resp := &model.Response{
+			Message: model.Message{
+				Status:    model.StatusNACK,
+				MessageID: msgID(ctx),
+				Error:     becknErr,
+			},
+		}
+		data, _ = json.Marshal(resp)
+	} else {
+		resp := &preV2Response{
+			Message: preV2Message{
+				Ack:   preV2Ack{Status: model.StatusNACK},
+				Error: becknErr,
+			},
+		}
+		data, _ = json.Marshal(resp)
+	}
+	return data
+}
+
+// nackBytes returns the NACK response body that sendNack would write for the
+// given error, without actually writing it. Use this to sign the body before
+// it is flushed to the wire (see stdHandler.signNackResponse).
+func nackBytes(ctx context.Context, err error) []byte {
+	becknErr, _ := nackBecknError(ctx, err)
+	return nackBodyBytes(ctx, becknErr)
+}
+
+// nack sends a NACK response with an error payload.
+// For context.version "2.0.0":
+//
+//	{"message":{"status":"NACK","messageId":"<uuid>","error":{...}}}
+//
+// All other versions:
+//
+//	{"message":{"ack":{"status":"NACK"},"error":{...}}}
+func nack(ctx context.Context, w http.ResponseWriter, becknErr *model.Error, status int) {
+	data := nackBodyBytes(ctx, becknErr)
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(status)
+	if _, er := w.Write(data); er != nil {
+		log.Debugf(ctx, "Error writing response: %v, MessageID: %s", er, ctx.Value(model.ContextKeyMsgID))
+		http.Error(w, fmt.Sprintf("Internal server error, MessageID: %s", ctx.Value(model.ContextKeyMsgID)), http.StatusInternalServerError)
+	}
+}
+
+// sendNack maps err to its Beckn error type and sends an appropriate NACK response.
+func sendNack(ctx context.Context, w http.ResponseWriter, err error) {
+	becknErr, status := nackBecknError(ctx, err)
+	nack(ctx, w, becknErr, status)
+}
+
+// isAtLeastV2 reports whether the request uses Beckn protocol v2.0.0 or later.
+func isAtLeastV2(ctx context.Context) bool {
+	v, _ := ctx.Value(model.ContextKeyProtocolVersion).(string)
+	return model.IsAtLeastV2(v)
+}
+
+// msgID returns the message ID stored in the context, or empty string if absent.
+func msgID(ctx context.Context) string {
+	v, _ := ctx.Value(model.ContextKeyMsgID).(string)
+	return v
+}
+
+// internalServerError generates a Beckn internal-server-error payload.
+func internalServerError(ctx context.Context) *model.Error {
+	return &model.Error{
+		Code:    http.StatusText(http.StatusInternalServerError),
+		Message: fmt.Sprintf("Internal server error, MessageID: %s", ctx.Value(model.ContextKeyMsgID)),
+	}
+}
+
+// ---------------------------------------------------------------------------
+// ackSignerStep — signs synchronous responses on the Receiver side
 // ---------------------------------------------------------------------------
 
 // ackSignerStep signs the synchronous Ack response per NFH-004 §3.4 and sets
@@ -72,7 +228,7 @@ func (a *ackSignerStep) signBodyAndSetHeader(ctx *model.StepContext, body []byte
 // RunOnResponse signs the Ack response and sets the Signature header.
 //
 // resp is nil on the publisher path: ONIX controls the ACK body, so the digest
-// is computed over the deterministic body that SendAck will write.
+// is computed over the deterministic body that sendAck will write.
 // resp is non-nil on the URL-routing path: the body comes from the upstream
 // app via ReverseProxy, so the digest covers the actual bytes the caller
 // receives. In both cases the Signature header value is identical in structure.
@@ -86,7 +242,6 @@ func (a *ackSignerStep) RunOnResponse(ctx *model.StepContext, resp *http.Respons
 	if !model.IsAtLeastV2(ctx.ProtocolVersion) {
 		return nil
 	}
-
 	if len(ctx.SubID) == 0 {
 		return model.NewBadReqErr(fmt.Errorf("subscriberID not set"))
 	}
@@ -116,7 +271,7 @@ func (a *ackSignerStep) RunOnResponse(ctx *model.StepContext, resp *http.Respons
 	}
 
 	// Publisher / no-route path: ONIX writes the ACK — build the deterministic
-	// body that SendAck will write so the digest matches.
+	// body that sendAck will write so the digest matches.
 	ackBody, err = buildAckBody(ctx.ProtocolVersion, ctx.MessageID)
 	if err != nil {
 		return fmt.Errorf("ackSigner: failed to build ack body: %w", err)
@@ -147,7 +302,7 @@ func buildSignatureHeader(subID, keyID string, createdAt, validTill int64, signa
 }
 
 // ---------------------------------------------------------------------------
-// validateAckSignatureStep
+// validateAckSignatureStep — verifies synchronous response signatures on the Caller side
 // ---------------------------------------------------------------------------
 
 // validateAckSignatureStep verifies the Signature response header on every
@@ -210,7 +365,6 @@ func newValidateAckSignatureStep(sv definition.SignValidator, km definition.KeyM
 //   - protocol version < 2.0.0 — pre-v2 flows are unaffected
 func (v *validateAckSignatureStep) RunOnResponse(ctx *model.StepContext, resp *http.Response) error {
 	if resp == nil {
-		// Publisher path — the ACK is sent asynchronously; no Signature header.
 		return nil
 	}
 	if !model.IsAtLeastV2(ctx.ProtocolVersion) {
@@ -219,8 +373,6 @@ func (v *validateAckSignatureStep) RunOnResponse(ctx *model.StepContext, resp *h
 
 	sigHeader := resp.Header.Get("Signature")
 	if sigHeader == "" {
-		// Missing Signature on any response (including error responses that the
-		// peer's ONIX hasn't yet been upgraded to sign) → degrade, not reject.
 		log.Warnf(ctx, "validateAckSign: missing Signature response header on v2 response (status=%d) — degraded trust", resp.StatusCode)
 		return nil
 	}
@@ -230,7 +382,6 @@ func (v *validateAckSignatureStep) RunOnResponse(ctx *model.StepContext, resp *h
 	// so the header value is the one sent to the upstream BPP/BAP.
 	outboundAuth := extractAuthSignature(ctx.Request.Header.Get(model.AuthHeaderSubscriber))
 
-	// Read and restore the ACK body for digest computation.
 	ackBody, err := io.ReadAll(resp.Body)
 	if err != nil {
 		log.Warnf(ctx, "validateAckSign: failed to read response body: %v — degraded trust", err)
@@ -238,14 +389,12 @@ func (v *validateAckSignatureStep) RunOnResponse(ctx *model.StepContext, resp *h
 	}
 	resp.Body = io.NopCloser(bytes.NewReader(ackBody))
 
-	// Parse the keyId from the Signature header to identify the signer.
 	parsed, err := parseHeader(sigHeader)
 	if err != nil {
 		log.Warnf(ctx, "validateAckSign: failed to parse Signature header keyId: %v — degraded trust", err)
 		return nil
 	}
 
-	// Look up the signer's public key from the registry via KeyManager.
 	publicKey, _, err := v.km.LookupNPKeys(ctx, parsed.SubscriberID, parsed.UniqueID)
 	if err != nil {
 		log.Warnf(ctx, "validateAckSign: failed to look up public key for %s: %v — degraded trust", parsed.SubscriberID, err)
